change type of SPIR-V code size to size_t

+++ src/include/vuh/program.hpp
@@ -8,7 +8,7 @@
 #include <vulkan/vulkan.hpp>
 
 #include <array>
-#include <cstddef>
+#include <stdint.h>
 #include <tuple>
 #include <utility>
 
@@ -147,13 +147,13 @@ namespace vuh {
 			            , const std::vector<char>& code  ///< actual binary SPIR-V shader code
 			            , vk::ShaderModuleCreateFlags flags={}
 			            )
-				: ProgramBase(device, uint32_t(code.size()),
+				: ProgramBase(device, code.size(),
 				              reinterpret_cast<const uint32_t*>(code.data()))
 			{}
 
 			/// Construct object using given a vuh::Device a SPIR-V shader code as plain array.
 			ProgramBase(vuh::Device& device              ///< device used to run the code
-			            , uint32_t size                  ///< size of binary shader code
+			            , size_t size                    ///< size of binary shader code
 			            , const uint32_t* code           ///< actual binary SPIR-V shader code
 			            , vk::ShaderModuleCreateFlags flags={}
 			            )
@@ -313,7 +313,7 @@ namespace vuh {
 			{}
 
 			/// Construct object using given a vuh::Device a SPIR-V shader code in plain array.
-			SpecsBase(Device& device, uint32_t size, const uint32_t* code, vk::ShaderModuleCreateFlags f={})
+			SpecsBase(Device& device, size_t size, const uint32_t* code, vk::ShaderModuleCreateFlags f={})
 			   : ProgramBase(device, size, code, f)
 			{}
 
@@ -349,7 +349,7 @@ namespace vuh {
 			{}
 
 			/// Construct object using given a vuh::Device a SPIR-V shader code in plain array.
-			SpecsBase(Device& device, uint32_t size, const uint32_t* code, vk::ShaderModuleCreateFlags f={})
+			SpecsBase(Device& device, size_t size, const uint32_t* code, vk::ShaderModuleCreateFlags f={})
 			   : ProgramBase(device, size, code, f)
 			{}
 
@@ -391,7 +391,7 @@ namespace vuh {
 		{}
 
 		/// Initialize program on a device from binary SPIR-V code in plain array
-		Program(vuh::Device& device, uint32_t size, const uint32_t* code
+		Program(vuh::Device& device, size_t size, const uint32_t* code
 		        , vk::ShaderModuleCreateFlags flags={}
 		        )
 		   : Base(device, size, code, flags)
@@ -492,7 +492,7 @@ namespace vuh {
 		{}
 
 		/// Initialize program on a device from binary SPIR-V code in plain array
-		Program(vuh::Device& device, uint32_t size, const uint32_t* code
+		Program(vuh::Device& device, size_t size, const uint32_t* code
 		        , vk::ShaderModuleCreateFlags flags={}
 		        )
 		   : Base(device, size, code, flags)
